removed installer files
I removed the installer files so I wouldnt have to upload them to github, and put download.txt files in their place with their download urls.

I have a script that can grab them and output them based on filenames contained in the urls, but some urls don't contain the filenames. I have a workaround where I can just add the desired filenames to the download.txt file, but there may be a better way to automate the population of installer files.

If the app is licensed or requires purchase/registration, there's a note about that in the Files directory of the apps PSADT folder.